Split out js event
